+++ CHANGELOG.rst
@@ -14,5 +14,6 @@ Fixed
 ~~~~~
 
 * (`#1`_) Check that Python can interpret ``HISTTIMEFORMAT``.
+* Do not assume ``HISTTIMEFORMAT`` evaluates to a fixed-length string.
 
 .. _#1: https://github.com/benwebber/duiker/issues/1